+++ tests/helpers/directives/TestDirective.ts
@@ -1,9 +1,9 @@
 import { MapperKind, getDirective, mapSchema } from "@graphql-tools/utils";
 import {
   DirectiveLocation,
+  type GraphQLArgumentConfig,
   GraphQLDirective,
   type GraphQLFieldConfig,
-  type GraphQLFieldExtensions,
   type GraphQLInputFieldConfig,
   GraphQLInputObjectType,
   type GraphQLInputObjectTypeConfig,
@@ -14,19 +14,23 @@ import {
   type GraphQLSchema,
 } from "graphql";
 
-function mapConfig(
-  config:
+function mapConfig<
+  TConfig extends
     | GraphQLFieldConfig<any, any, any>
     | GraphQLObjectTypeConfig<any, any>
     | GraphQLInterfaceTypeConfig<any, any>
     | GraphQLInputObjectTypeConfig
-    | GraphQLInputFieldConfig,
-) {
-  // eslint-disable-next-line no-param-reassign
-  config.extensions ??= {};
-  // eslint-disable-next-line no-param-reassign
-  (config.extensions as GraphQLFieldExtensions<any, any, any>).TypeGraphQL = {
-    isMappedByDirective: true,
+    | GraphQLInputFieldConfig
+    | GraphQLArgumentConfig,
+>(config: TConfig) {
+  return {
+    ...config,
+    extensions: {
+      ...config.extensions,
+      TypeGraphQL: {
+        isMappedByDirective: true,
+      },
+    },
   };
 }
 
@@ -38,6 +42,7 @@ export const testDirective = new GraphQLDirective({
     DirectiveLocation.INPUT_OBJECT,
     DirectiveLocation.INPUT_FIELD_DEFINITION,
     DirectiveLocation.INTERFACE,
+    DirectiveLocation.ARGUMENT_DEFINITION,
   ],
 });
 
@@ -47,15 +52,14 @@ export function testDirectiveTransformer(schema: GraphQLSchema): GraphQLSchema {
       const testDirectiveConfig = getDirective(schema, typeInfo, testDirective.name)?.[0];
       if (testDirectiveConfig) {
         const config = typeInfo.toConfig();
-        mapConfig(config);
-        return new GraphQLObjectType(config);
+        return new GraphQLObjectType(mapConfig(config));
       }
       return typeInfo;
     },
     [MapperKind.OBJECT_FIELD]: fieldConfig => {
       const testDirectiveConfig = getDirective(schema, fieldConfig, testDirective.name)?.[0];
       if (testDirectiveConfig) {
-        mapConfig(fieldConfig);
+        return mapConfig(fieldConfig);
       }
       return fieldConfig;
     },
@@ -63,15 +67,14 @@ export function testDirectiveTransformer(schema: GraphQLSchema): GraphQLSchema {
       const testDirectiveConfig = getDirective(schema, interfaceConfig, testDirective.name)?.[0];
       if (testDirectiveConfig) {
         const config = interfaceConfig.toConfig();
-        mapConfig(config);
-        return new GraphQLInterfaceType(config);
+        return new GraphQLInterfaceType(mapConfig(config));
       }
       return interfaceConfig;
     },
     [MapperKind.INTERFACE_FIELD]: fieldConfig => {
       const testDirectiveConfig = getDirective(schema, fieldConfig, testDirective.name)?.[0];
       if (testDirectiveConfig) {
-        mapConfig(fieldConfig);
+        return mapConfig(fieldConfig);
       }
       return fieldConfig;
     },
@@ -79,17 +82,23 @@ export function testDirectiveTransformer(schema: GraphQLSchema): GraphQLSchema {
       const testDirectiveConfig = getDirective(schema, typeInfo, testDirective.name)?.[0];
       if (testDirectiveConfig) {
         const config = typeInfo.toConfig();
-        mapConfig(config);
-        return new GraphQLInputObjectType(config);
+        return new GraphQLInputObjectType(mapConfig(config));
       }
       return typeInfo;
     },
     [MapperKind.INPUT_OBJECT_FIELD]: fieldConfig => {
       const testDirectiveConfig = getDirective(schema, fieldConfig, testDirective.name)?.[0];
       if (testDirectiveConfig) {
-        mapConfig(fieldConfig);
+        return mapConfig(fieldConfig);
       }
       return fieldConfig;
+    },
+    [MapperKind.ARGUMENT]: argumentConfig => {
+      const testDirectiveConfig = getDirective(schema, argumentConfig, testDirective.name)?.[0];
+      if (testDirectiveConfig) {
+        return mapConfig(argumentConfig);
+      }
+      return argumentConfig;
     },
   });
 }